requirements.txt
pymupdf
langchain
langchain-openai
qdrant-client
openai
python-dotenv
tqdm